requirements.txt
pytelegrambotapi
langchain
openai
huggingface_hub
pinecone-client
pydantic[dotenv]
transformers

